JETTY-609 work in progress
Refactor the HttpClient security conversations.

HttpDestinations now have a repository of cookies and authentications

The concept of HttpConversation is now just a HttpEventListener

HttpDestinations insert a security listener to intercept events (if there is a realm).
If that listener gets a 401 unauthorized response, it stops delegating events to the 
wrapped listener (normally the exchange) and organizes to do a retry.  Before doing
the retry the HttpDestination is updated with an Authentication (not completely implemented yet).

The SecurityResolver is now a SecurityListener




git-svn-id: https://svn.codehaus.org/jetty/jetty/trunk/modules@3004 c7fae8f8-2506-0410-b6b2-ad6cf81c2737